refactors sound toggle prefs away from legacy toggles
