+++ code/__DEFINES/preferences.dm
@@ -3,15 +3,9 @@
 // !!! DO NOT ADD ANY NEW ONES HERE !!!
 // Use `/datum/preference/toggle` instead.
 #define SOUND_ADMINHELP (1<<0)
-#define SOUND_MIDI (1<<1)
-#define SOUND_AMBIENCE (1<<2)
-#define SOUND_LOBBY (1<<3)
 #define MEMBER_PUBLIC (1<<4)
-#define SOUND_INSTRUMENTS (1<<7)
-#define SOUND_SHIP_AMBIENCE (1<<8)
 #define SOUND_PRAYERS (1<<9)
 #define ANNOUNCE_LOGIN (1<<10)
-#define SOUND_ANNOUNCEMENTS (1<<11)
 #define DISABLE_DEATHRATTLE (1<<12)
 #define DISABLE_ARRIVALRATTLE (1<<13)
 #define COMBOHUD_LIGHTING (1<<14)
@@ -20,12 +14,10 @@
 #define DEADMIN_POSITION_HEAD (1<<17)
 #define DEADMIN_POSITION_SECURITY (1<<18)
 #define DEADMIN_POSITION_SILICON (1<<19)
-#define SOUND_ENDOFROUND (1<<20)
 #define ADMIN_IGNORE_CULT_GHOST (1<<21)
-#define SOUND_COMBATMODE (1<<22)
 #define SPLIT_ADMIN_TABS (1<<23)
 
-#define TOGGLES_DEFAULT (SOUND_ADMINHELP|SOUND_MIDI|SOUND_AMBIENCE|SOUND_LOBBY|SOUND_ENDOFROUND|MEMBER_PUBLIC|SOUND_INSTRUMENTS|SOUND_SHIP_AMBIENCE|SOUND_PRAYERS|SOUND_ANNOUNCEMENTS|SOUND_COMBATMODE)
+#define TOGGLES_DEFAULT (SOUND_ADMINHELP|MEMBER_PUBLIC|SOUND_PRAYERS)
 
 // Legacy chat toggles.
 // !!! DO NOT ADD ANY NEW ONES HERE !!!
