+++ code/modules/instruments/songs/play_legacy.dm
@@ -85,7 +85,7 @@
 		if(player && HAS_TRAIT(player, TRAIT_MUSICIAN) && isliving(M))
 			var/mob/living/L = M
 			L.apply_status_effect(/datum/status_effect/good_music)
-		if(!(M?.client?.prefs?.toggles & SOUND_INSTRUMENTS))
+		if(!(M?.client?.prefs.read_preference(/datum/preference/toggle/sound_instruments)))
 			continue
 		M.playsound_local(source, null, volume * using_instrument.volume_multiplier, sound_to_use = music_played)
 		// Could do environment and echo later but not for now
