+++ code/modules/mob/living/silicon/ai/ai_say.dm
@@ -153,7 +153,7 @@
 					stack_trace("[player_mob] ([player_mob.ckey]) has null prefs, which shouldn't be possible!")
 					continue
 
-				if(player_mob.can_hear() && (player_mob.client?.prefs.toggles & SOUND_ANNOUNCEMENTS))
+				if(player_mob.can_hear() && (player_mob.client?.prefs.read_preference(/datum/preference/toggle/sound_announcements)))
 					var/turf/T = get_turf(player_mob)
 					if(T.z == z_level)
 						SEND_SOUND(player_mob, voice)
